+++ src/components/detalle-full-page/detalle-full-page-components/SeccionPropiedad.tsx
@@ -33,24 +33,24 @@ export const SeccionPropiedad = ({propiedad}: SeccionPropiedadProps) => {
       </div>
 
       {/* Detalles enumerados */}
-      <h2 className="font-bold text-black text-xl my-8"><span className="text-foreground">|&nbsp;</span>Detalles</h2>
-      <DetallesGrid
-        propiedad={ propiedad }
-        detalles={[
-          { descripcion: 'Estilo',                  clave: 'estilo' },
-          { descripcion: 'Nro. de plantas',         clave: 'plantas'},
-          { descripcion: 'Dormitorios',             clave: 'dormitorios'},
-          { descripcion: 'Dormitorios en suite',    clave: 'dormitorio_suite'},
-          { descripcion: 'Estado',                  clave: 'estado'},
-          { descripcion: 'Antiguedad',              clave: 'antiguedad'},
-          { descripcion: 'Lote',                    clave: 'lote'},
-          { descripcion: 'Tipo de zona',            clave: 'tipo_zona'},
-          { descripcion: 'Tipo de calefacción',     clave: 'calefaccion'},
-          { descripcion: 'Aberturas',               clave: 'aberturas'},
-        ]}/>
+      <h2 className="font-bold text-black text-xl mt-8 mb-4"><span className="text-foreground">|&nbsp;</span>Detalles</h2>
+        <DetallesGrid
+          propiedad={ propiedad }
+          detalles={[
+            { descripcion: 'Estilo',                  clave: 'estilo' },
+            { descripcion: 'Nro. de plantas',         clave: 'plantas'},
+            { descripcion: 'Dormitorios',             clave: 'dormitorios'},
+            { descripcion: 'Dormitorios en suite',    clave: 'dormitorio_suite'},
+            { descripcion: 'Estado',                  clave: 'estado'},
+            { descripcion: 'Antiguedad',              clave: 'antiguedad'},
+            { descripcion: 'Lote',                    clave: 'lote'},
+            { descripcion: 'Tipo de zona',            clave: 'tipo_zona'},
+            { descripcion: 'Tipo de calefacción',     clave: 'calefaccion'},
+            { descripcion: 'Aberturas',               clave: 'aberturas'},
+          ]}/>
 
       {/* Descripcion */}
-      <h2 className="font-bold text-black text-xl my-4"><span className="text-foreground">|&nbsp;</span>Descripción</h2>
+      <h2 className="font-bold text-black text-xl mt-8 mb-4"><span className="text-foreground">|&nbsp;</span>Descripción</h2>
     
       {/* Descripcion larga */}
       <p className={`${secondaryFont.className} text-black text-lg`}>{ descripcion_larga }</p>
